+++ pkg/querier/handler.go
@@ -125,10 +125,16 @@ func (h *Handler) Do(ctx context.Context, req queryrangebase.Request) (queryrang
 		if err != nil {
 			return nil, err
 		}
-
 		return &queryrange.QueryPatternsResponse{
 			Response: result,
 		}, nil
+	case *queryrange.DetectedLabelsRequest:
+		result, err := h.api.DetectedLabelsHandler(ctx, &concrete.DetectedLabelsRequest)
+		if err != nil {
+			return nil, err
+		}
+
+		return &queryrange.DetectedLabelsResponse{Response: result}, nil
 	default:
 		return nil, fmt.Errorf("unsupported query type %T", req)
 	}