+++ pkg/querier/queryrange/roundtrip.go
@@ -246,22 +246,23 @@ func NewMiddleware(
 			statsRT          = indexStatsTripperware.Wrap(next)
 			seriesVolumeRT   = seriesVolumeTripperware.Wrap(next)
 			detectedFieldsRT = next // TODO(twhitney): add middlewares for detected fields
+			detectedLabelsRT = next // TODO(shantanu): add middlewares
 		)
 
-		return newRoundTripper(log, next, limitedRT, logFilterRT, metricRT, seriesRT, labelsRT, instantRT, statsRT, seriesVolumeRT, detectedFieldsRT, limits)
+		return newRoundTripper(log, next, limitedRT, logFilterRT, metricRT, seriesRT, labelsRT, instantRT, statsRT, seriesVolumeRT, detectedFieldsRT, detectedLabelsRT, limits)
 	}), StopperWrapper{resultsCache, statsCache, volumeCache}, nil
 }
 
 type roundTripper struct {
 	logger log.Logger
 
-	next, limited, log, metric, series, labels, instantMetric, indexStats, seriesVolume, detectedFields base.Handler
+	next, limited, log, metric, series, labels, instantMetric, indexStats, seriesVolume, detectedFields, detectedLabels base.Handler
 
 	limits Limits
 }
 
 // newRoundTripper creates a new queryrange roundtripper
-func newRoundTripper(logger log.Logger, next, limited, log, metric, series, labels, instantMetric, indexStats, seriesVolume, detectedFields base.Handler, limits Limits) roundTripper {
+func newRoundTripper(logger log.Logger, next, limited, log, metric, series, labels, instantMetric, indexStats, seriesVolume, detectedFields, detectedLabels base.Handler, limits Limits) roundTripper {
 	return roundTripper{
 		logger:         logger,
 		limited:        limited,
@@ -274,6 +275,7 @@ func newRoundTripper(logger log.Logger, next, limited, log, metric, series, labe
 		indexStats:     indexStats,
 		seriesVolume:   seriesVolume,
 		detectedFields: detectedFields,
+		detectedLabels: detectedLabels,
 		next:           next,
 	}
 }
@@ -378,6 +380,7 @@ func (r roundTripper) Do(ctx context.Context, req base.Request) (base.Response, 
 		)
 
 		return r.detectedFields.Do(ctx, req)
+	// TODO(shantanu): Add DetectedLabels
 	default:
 		return r.next.Do(ctx, req)
 	}
@@ -413,6 +416,7 @@ const (
 	IndexShardsOp    = "index_shards"
 	DetectedFieldsOp = "detected_fields"
 	PatternsQueryOp  = "patterns"
+	DetectedLabelsOp = "detected_labels"
 )
 
 func getOperation(path string) string {
@@ -437,6 +441,8 @@ func getOperation(path string) string {
 		return DetectedFieldsOp
 	case path == "/loki/api/v1/patterns":
 		return PatternsQueryOp
+	case path == "/loki/api/v1/detected_labels":
+		return DetectedLabelsOp
 	default:
 		return ""
 	}