+++ pkg/querier/queryrange/stats.go
@@ -38,6 +38,7 @@ const (
 	queryTypeShards         = "shards"
 	queryTypeDetectedFields = "detected_fields"
 	queryTypeQueryPatterns  = "patterns"
+	queryTypeDetectedLabels = "detected_labels"
 )
 
 var (
@@ -65,6 +66,8 @@ func recordQueryMetrics(data *queryData) {
 		logql.RecordVolumeQueryMetrics(data.ctx, logger, data.params.Start(), data.params.End(), data.params.QueryString(), data.params.Limit(), data.params.Step(), data.status, *data.statistics)
 	case queryTypeDetectedFields:
 		logql.RecordDetectedFieldsQueryMetrics(data.ctx, logger, data.params.Start(), data.params.End(), data.params.QueryString(), data.status, *data.statistics)
+	case queryTypeDetectedLabels:
+		logql.RecordDetectedLabelsQueryMetrics(data.ctx, logger, data.params.Start(), data.params.End(), data.params.QueryString(), data.status, *data.statistics)
 	default:
 		level.Error(logger).Log("msg", "failed to record query metrics", "err", fmt.Errorf("expected one of the *LokiRequest, *LokiInstantRequest, *LokiSeriesRequest, *LokiLabelNamesRequest, got %s", data.queryType))
 	}