+++ pkg/util/marshal/marshal.go
@@ -229,3 +229,13 @@ func WriteQueryPatternsResponseJSON(r *logproto.QueryPatternsResponse, w io.Writ
 	s.WriteRaw("\n")
 	return s.Flush()
 }
+
+// WriteDetectedLabelsResponseJSON marshals a logproto.DetectedLabelsResponse to JSON and then
+// writes it to the provided io.Writer.
+func WriteDetectedLabelsResponseJSON(r *logproto.DetectedLabelsResponse, w io.Writer) error {
+	s := jsoniter.ConfigFastest.BorrowStream(w)
+	defer jsoniter.ConfigFastest.ReturnStream(s)
+	s.WriteVal(r)
+	s.WriteRaw("\n")
+	return s.Flush()
+}